Tried to get mouse-events (but could not)

+++ deprecated.go
@@ -6,7 +6,7 @@ var OnWindowResize func(w, h uint)
 // Deprecated: use All()
 // Get all keyboard-event.
 func Full() (code rune, scan uint16, shift uint32) {
-	var flag uintptr = 0
+	var flag uint32 = 0
 	if OnWindowResize != nil {
 		flag = ENABLE_WINDOW_INPUT
 	}
